+++ planner/include/planner/util.h
@@ -24,4 +24,40 @@ namespace util
         geometry_msgs::Vector3 dimension;
         geometry_msgs::Pose pose;
     } CollisionGeometry;
+
+    inline bool operator==(const CollisionGeometry& lhs, const CollisionGeometry& rhs)
+    {
+        return lhs.name.compare(rhs.name) == 0 &&
+            lhs.min == rhs.min &&
+            lhs.max == rhs.max &&
+            lhs.pose == rhs.pose &&
+            lhs.dimension == rhs.dimension;
+    }
+
+    inline bool operator!=(const CollisionGeometry& lhs, const CollisionGeometry& rhs)
+    {
+        return !(lhs == rhs);
+    }
+
+    template <typename T>
+    T clamp(T val, T lower, T upper)
+    {
+        if (val < lower) return lower;
+        if (val > upper) return upper;
+        return val;
+    }
+    
+    // mainly for use with floating point vectors
+    template <typename T>
+    bool approxEqual(std::vector<T> vec1, std::vector<T> vec2, T max_error)
+    {
+        assert(vec1.size() == vec2.size() && "Size mismatch! Unable to check if Vectors are approximately equal.");
+
+        for (int i = 0; i < vec1.size(); i++)
+        {
+            if (std::abs(vec1[i] - vec2[i]) > max_error) return false;
+        }
+
+        return true;
+    }
 }